--- src/renderer/src/components/D-Info.vue
@@ -1,11 +1,0 @@
-<template>
-    <div class="flex items-center justify-start gap-4">
-        <header class="self-start mt-4">
-            <slot name="header"></slot>
-        </header>
-
-        <main>
-            <slot name="content"></slot>
-        </main>
-    </div>
-</template>
