+++ src/renderer/src/components/EventSnackbar.vue
