+++ src/renderer/src/views/group/Group.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script setup lang="ts">
-import Card from '@/components/card/Card.vue'
+import Card from '@/components/Card.vue'
 import { useTags } from '@/hooks/useTags'
 import { useTaskStore } from '@/store/useTaskStore'
 import { useRouter } from 'vue-router'
