+++ src/renderer/src/views/home/lib/Result.vue
@@ -83,7 +83,7 @@
 </template>
 
 <script setup lang="ts">
-import Task from '@/components/task/Task.vue'
+import Task from '@/components/Task.vue'
 import { Item } from '@/hooks/useItem'
 import { TASKS_TYPE, useTaskStore } from '@/store/useTaskStore'
 import { useRouter } from 'vue-router'
